pyproject.toml
[tool.poetry]
name = "niva-ai-backend"
version = "0.1.0"
description = ""
authors = ["govardhanar06@gmail.com"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "5.2.1"
django-cors-headers = "4.7.0"
djangorestframework = "3.16.0"
aiohttp = "3.11.18"
pipecat-ai = {extras = ["daily", "google"], version = "^0.0.74"}
celery = "5.5.2"
redis = "6.1.0"
psycopg2-binary = "2.9.10"
channels = "4.2.2"
django-phonenumber-field = "8.1.0"
phonenumbers = "8.13.30"
django-celery-beat = "2.8.1"
drf-yasg = "1.21.10"
django-redis = "5.4.0"
hiredis = "3.1.1"
channels-redis = "4.2.1"
pgvector = "0.4.1"
scikit-learn = "^1.6.1"
unstructured-client = "0.32.0"
pydantic = "2.10.6"
google-cloud-storage = "3.1.0"
langchain = "^0.3.25"
langchain-community = "0.3.10"
bs4 = "^0.0.2"
filelock = "^3.18.0"
jsonschema = "4.24.0"
PyMuPDF = "1.26.0"
pipecat-ai-flows = "0.0.18"
aiofiles = "^24.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "7.4.4"
black = "23.12.1"
isort = "5.13.2"
flake8 = "6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"